Initial port to PyGI, the Python bindings for GObject Introspection
With these changes, most of the test suite passes. The remaining problem is
that the messages our callbacks receive from GStreamer is None. See
https://bugs.launchpad.net/pygobject/+bug/873712 for details.

When the remaining issue is resolved, most likely with the transition from
GStreamer 0.10 to 0.11/1.0, this will remove our dependency on the Python
bindings for GStreamer, GObject, and libindicate. Replacing this pile of
manually maintained Python 2.x bindings with PyGI, which works under Python 3
as well, is a huge step towards bringing Mopidy to Python 3.

PyGI may also remove our dependency on the dbus binding, but this has not been
adressed in this commit as e.g.
https://bugzilla.gnome.org/show_bug.cgi?id=656330 hints that it is not possible
to port DBus servers to PyGI yet.