Removed autoaudiosink's device selection.
The way this code was testing devices locked the element to using the wrong
device. The incorrect device had a max volume of 39 on the Master track, really
making accurate volume changes impossible.

Instead of trying to make any guesses about this I'm leaving it to the element
to have sensible defaults.

Code will also ensure that it returns a newly created copy of the mixer, not
one we have already used.