Hack to fix random test failure
With Pykka 0.16, test_status_method_when_playing_contains_time_with_length
fails now and then because play_time_started is not initialized before it is
used as an int.

I'm allowing myself to fix this in the simplest way possible instead of
tracking the issue down, since I'm already working on a refactor of the time
position code.