Store active appsrc and refuse data when it is not set.
We use the new source flag and the about to finish flags to set and unset the
current appsrc. In emit data we now return false if the appsrc is not set.

Also note that we need to use b'' for Gstreamer properties as it can't convert
unicode to the correct type. I also added the signal disconnect code for about
to finish.