mpd: Limit use of fake tracks in 'find` responses
If searching for exact artist, don't include fake artist tracks.
If searching for exact album, don't include fake album tracks.

This makes sure that ncmpcpp's media library doesn't include the magic
artist-track in an artist's album listing, and that it doesn't include the
magic album-track in an album's track listing.