Add timestamp and duration to all spotify buffers.
This fixes the issue where pausing playback would show the time of the last
timestamped buffer instead of the current time. We also make sure to reset the
time when we start a new track. This was done by overriding the play method on
the session manager as it is also used for pausing, resuming and stopping.

Ideally this should probably be reworked to avoid the gst import in
mopidy.backends.spotify.playback, but for now this should do.