docs: Bind console functions in JS examples
For e.g. ``console.log`` to be used as a callback, it must be bound to the
``console`` object first: ``console.log.bind(console)``. If not, this will
cause "Illegal invocation" errors in WebKit browsers.

This change updates all our Mopidy.js examples to bind ``console`` functions
before they are used as callbacks.

Fixes #298.