Change docs license to Apache License, Version 2.0
I've looked through all contributions to the docs. Any contributions not by
Thomas Adamcik or myself are trivial (e.g. typo fixes) or have mostly been
replaced by later changes. Thus I'm comfortable with changing the license with
the agreement of Thomas Adamcik.

By changing the documentation license to be the same as the source code
license, we avoid ambiguities around the license of e.g. docstrings which are
written as part of the source code but are included in the generated docs.