local: Temporary workaround for issue #527
Adds a fallback to `Track(uri=uri` when track lookup fails for playlists. This
means we can at least load metadata less tracks giving users functioning
playlists, instead of only supporting `local:track:...` style URIs.

Issue is not fixed, but this is sufficient to reduce priority until we get to
the larger planed refactor for this and other core API issues.