Document workaround example for #492
The example is using an Icecast fallback stream that plays a silent MP3 file
when Mopidy sends the shutdown signal.

Tested and found working, however had minor issues with the client
(mpg123) playing the stream was getting out of sync when rapidly changing
songs.
Looking at the Icecast logs, it revealed that the switch between the mopidy and
the fallback stream happened immediately, so the issue is probably with the
client.