docs: Move images out of the _static dir
Sphinx copies the images to _images when the docs are built. Thus the
mopidy-docs debian package contains two copies of all images; one in _static
and one in _images. By keeping the images next to the source documents we only
get the _images copy included in the Debian package, reducing the package size
by 1.5MB.