Pause gstreamer when buffering
When the `playbin2` pipeline is in `PLAYING` state and its buffer is empty the
pipeline will consume arriving data immediately. If the source is an HTTP
stream this leads to choppy playback. To fix this we pause playing when the
buffer is nearly empty and wait for it to fill up until we resume playing. This
approach is recommended in the gstreamer manual [1].

We might want to extract the hard-coded 10% mark and make it configurable.

[1]: http://gstreamer.freedesktop.org/data/doc/gstreamer/head/manual/html/chapter-buffering.html