zeroconf: Fix intermittent dbus/avahi exception
This fixes an issue where I sometimes would get an error from dbus
'Unable to guess signature from an empty list'. After some digging
and checking the avahi dbus specs I found they expect the text list
to have a signature of 'aay' (an array of arrays containing bytes).
So instead of using python lists we now use a 'typed' dbus array.

It is not clear to me why this is a heisenbug, but this fix does
seem to make it go away.