core: Reduce stream metadata to just the title
The API I really want for this to support regular tracks, stream updates and
dynamic playlists is still unclear to me. As such I'm taking the KISS approach
and reducing this to just the stream title and nothing else.

If all goes as planed this will be replaced by playback_track_changed(tlid, ref)
style events and other improvements in a later version.