core: Add a [core] config section
I deemed it better to make core an extension (that cannot be disabled)
rather than add a special case throughout the config handling to make
it possible to have config section that doesn't belong to an extension.

This change is needed for #997.

Until #997 is completed, Mopidy will complain that core has now config
schema (because it is empty) and claim that core is disabled. This of
course has no practical effects.