tests: fix test breakage due to promotion from int to long
In internals/path.py. there is a snippet of code that multiples mtime
for a file with 1000, and then casting it to `int`. This pseudo-casting
is pointless, as it only truncates to an integer (whole number), instead
of actuallly ensuring the type is an <int>.

>>> type(int(2**32))
<type 'long'>

Instead, fix the tests to look for (int, long).

This bug found on a 32-bit VM :)