stream: Ignore file protocol if Mopidy-File is enabled
If Mopidy-File is enabled it handles playback of file:// URIs.

Mopidy-Stream used to do this, but in Mopidy 1.1 we removed "file" from
the default value of the stream/protocols config. However, many users
upgrading to Mopidy 1.1 have set stream/protocols to include "file" in
their existing config, and thus Mopidy fails to start because both
backends tries to claim the "file" protocol.

Fixes #1248