listener: Kill off mopidy.listener.send_async
This is no longer needed as the plain send method makes sure to use tell to
queue actor message. Which has better performance, and avoids deadlocks.

A side effect of this is that assuming you have a core actor running and a
dummy audio in use audio events just work.