Docs: Add Mopidy-Jukepi to the list of webclients
jukePi has been formally converted in to a Mopidy extension and
submitted to PyPi. In the current documentation, it's listed in the
'Other Web Clients' section.

This change adds a screenshot and moves it to its own section in the web
extensions doc.

(cherry picked from commit 6b1707d120fb1b039576b2099e6fb266733eab2c)