Fix issues relating to removing the current tl_track from the tracklist.
1. Stop playing before clearing the tracklist
2. Stop playback on remove() if the track being removed is the current 
tl_track.

This has knock-on effects if consukme is enabled, because consume 
removes the current tl_track on track_playback_ended, which now makes 
playback stop, so

3. Change that to consume the *previous* track on track_playback_start
4. Consume the current tl_track on stop
5. Consume the current tl_track on end_of_stream, to catch the case 
where we've reached the end of the tracklist

And there's still a problem with streams:
On stopping a stream, before Mopidy can change state to STOPPED, 
GStreamer emits a buffering message with a percent of 0 - presumably 
because it empties its buffer. On receipt of this we change playback 
state to PAUSE. If the stream is then removed from the tracklist the 
playbin gets left in PAUSED and trying to start playback of a new track 
or stream simply unpauses the playbin and the removed stream keeps 
playing.
Checking that the target state is not NULL before acting on the 
buffering message catches the case where we've stopped the stream, but 
it doesn't work when switching from one stream to another with consume 
enabled. The only sure is to not act on a buffering message with a 
percent of 0