Fix various issues relating to removing the current tl_track from the tracklist. Doing this with the playback state at anything but STOPPED leaves the track stuck in GStreamer's playbin and it will continue playing.
1. Stop playing before clearing the tracklist
2. Stop playback on remove() if the track being removed is the current
tl_track.

This has knock-on effects if consume is enabled, because consume
removes the current tl_track on track_playback_ended, which now makes
playback stop, so

3. Change that to consume the *previous* track on track_playback_start
4. Consume the current tl_track after stop
5. Consume the current tl_track on end_of_stream, to catch the case
where we've reached the end of the tracklist

6. Fix a related issue with next() and play() if the track is not
playable where the unplaybale track was removed from the tracklist but
then used to look up next_track, which then returns the first track in
the tracklist

7. When stopping a stream, GStreamer sends a buffering message with a
percent of 0, which makes audio/actor.py change playback state to paused
when we want stopped. This affects us badly if consume is on because we
then remove the current track while it's paused. So we need to check the
audio target_state before we act on a buffering message to make sure
the target state is play.