config: Use text as the internal serialization format
The config value's own `encode()` method ensures that we use the
`surrogateescape` error handler, so that all bytes will survive the
conversion to text and back to bytes.

With this rewrite, only the `encode()` function and the two places that
outputs the config document, either to file or console, have to care
about "surrogateescape". Everything else is just concerned with text.

Fixes #1850 (partly)