audio: Fix logging of GStreamer warning/error
On Python 2, we used str(error) to get a native string representation,
aka bytes, and then used .decode() on that to get a unicode string.

On Python 3, the str(error) call returns Unicode, which doesn't have a
.decode() method, so the error/warning handler crashes.

I added ipdb tracing, provoked an error, and investigated the objects we
get from GStreamer:

- `error` is an GLib.GError which has a good str()/repr(), as well as
  .message, .domain, and .code attributes.
- `debug` is a string

There were no traces of bytes that needed decoding in the objects.

Fixes #1851