core: implement optional pagination for search API
This enables library providers to optionally implement pagination when
using the search API. The motivation for this is the fact that currently
`Mopidy-Local` only returns 100 results per search query, which means
that large parts of my local library are inaccessible by search. This PR
adds `limit` and `offset` keyword arguments to the search API that are
then passed to each backend when searching. However, in order to
preserve backwards compatibility, this is only done if the library
provider has the attribute `supports_pagination` and the value is
`True`. This partially addresses issue #733 but only for search.