Resolve issue with backend lists preparation
Previously when preparing sub-lists of backends (based on what they
support), we would still try to add the relevant backend to each list
even if an exception was thrown while fetching the relevant information.
This was problematic:

- If the very first backend encountered an issue, this code would also
  throw an exception because some variables would be undefined
- If any other backends encountered an issue, this code would make
  decisions based on the data obtained from a previous backend

Now we don't try to proceed with adding a backend to any sub-list if an
error occurs.

Also, the code used to construct the relevant error message in this
situation now uses an existing function to get the relevant backend's
classname.