Add core status function
In situations where it isn't practical to use websockets, the only
reliable way to retrieve the current status of the server is to poll to
the JSON-RPC endpoint. Unfortunately, there hasn't been a single unified
"get status of the server" function up until now. In order to determine
whether or not "repeat", "random", "single" and "consume" are enabled,
you would need to make four separate requests. Now, you will only need
to request a single endpoint.

The output was modelled on the existing "status" endpoint provided by
the MPD frontend. It includes a little more information out of the box,
and isn't aiming to be 100% compatible with what MPD provides. It is
expected that once this change has been released, an update can be made
to the MPD frontend to utilise this, rather than request all the
different functions itself.