Simplify Debian repo key installation
The `mopidy.gpg` file is now dearmored, removing the need for `gpg` to
be installed on the enduser's system.

The use of `tee` has been replaced by running `wget` as root. We already
run `wget` as root in the next step, so this isn't something we're
trying to avoid. The main benefit is to avoid using `tee`, which outputs
the binary contents of `mopidy.gpg` to the terminal too, risking that
many users will think something is broken.