Reencode test files to fix duration mismatch
Recently we've seen test runs on both Ubuntu devel and Arch fail because
the duration of `song1.mp3` did no longer match the number of
milliseconds in the test case. The commonality between these two systems
are probably that they both have GStreamer 1.22.

We could probably just assert that the duration is positive, but I
suspect this is caused by mismatch between metadata and actual duration
or simply that the file was encoded with a very old decoder.

In this PR, I've simply reencoded the `song1.mp3` and `song1.ogg` files
by running `ffmpeg -i old-file new-file` on the files, and then updated
the test case with the new durations as reported on my Ubuntu 22.04
system with GStreamer 1.20.

It would be great if someone could confirm that the test still passes on
GStreamer 1.22 before we merge this.

Fixes #2077