[#2202] Workaround for race conditions with recent versions of GStreamer (>= 1.25.x)
There are several issues with the current audio pipeline setup with
recent versions of GStreamer. And it seems that most of the issue can be
pinned down to:

1. Pipelines and sinks being reused in the same actor context

2. The `audio-sink` property being set at runtime

This causes resources associated to some remote extensions (tested with
mopidy-tidal, mopidy-jellyfin and mopidy-soundcloud) to freeze the
application (most likely because of deadlocks in the Gst stack) when
played.

This commit implements a workaround that:

1. Re-creates and reconfigures the `playbin` object every time that
   `set_uri` is called (emulating a clean context/state that you'd get
   by running a stand-alone `gst-launch-1.0` command)

2. Avoid re-initializing the sink from the `set_uri` context. After
   extensive testing, I have noticed that the deadlock occurs when the
   sink is created and the pipeline is already in a non-NULL state.

The latter requires further refinement - the downside of the current
approach is that it's not possible to control the volume through the
built-in mixer.